+++ src/utils.ts
@@ -59,7 +59,6 @@ export async function postDegreeOfSuccessMessage(degreeOfSuccess: DegreeOfSucces
         message = messageConfig.critFailure;
     }
     if (message !== '') {
-        console.debug(message);
         await ChatMessage.create({
             type: CONST.CHAT_MESSAGE_TYPES.ROLL,
             content: message.trimEnd(),